*** empty log message ***

+++ lib/crypt/rc4.c
@@ -38,7 +38,7 @@ void rc4_crypt(unsigned char *buffer_ptr, int buffer_len, rc4_key *key)
 	unsigned char y;
 	unsigned char* state;
 	unsigned char xorIndex;
-	short counter;              
+	int counter;              
 	
 	x = key->x;     
 	y = key->y;
